id type changes

+++ src/main/java/com/ftts/repository/RaceRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(path = "race")
-public interface RaceRepository extends JpaRepository<Race, Integer>  {
+public interface RaceRepository extends JpaRepository<Race, Long>  {
     Race getById(@Param("id") Integer id);
 
     @Query(value = "SELECT r from Race r WHERE raceName LIKE %:raceName%")
